Runs the flutter migrate and applies changes

+++ example/supporting_files/parameters/power_of_two.parameter.dart
@@ -5,6 +5,6 @@ class PowerOfTwoParameter extends CustomParameter<int> {
   PowerOfTwoParameter()
       : super('POW', RegExp(r'([0-9]+\^[0-9]+)', caseSensitive: true), (input) {
           final parts = input.split('^');
-          return pow(int.parse(parts[0]), int.parse(parts[1]));
+          return pow(int.parse(parts[0]), int.parse(parts[1])) as int;
         });
 }
